Rename planner plugin to deployment plugin
