+++ pkg/app/pipedv1/plugin/kubernetes/deployment/server.go
@@ -1,0 +1,72 @@
+// Copyright 2024 The PipeCD Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package deployment
+
+import (
+	"context"
+
+	"github.com/pipe-cd/pipecd/pkg/plugin/api/v1alpha1/deployment"
+	"github.com/pipe-cd/pipecd/pkg/regexpool"
+
+	"go.uber.org/zap"
+	"google.golang.org/grpc"
+)
+
+type DeploymentService struct {
+	deployment.UnimplementedDeploymentServiceServer
+
+	RegexPool *regexpool.Pool
+	Logger    *zap.Logger
+}
+
+// NewDeploymentService creates a new planService.
+func NewDeploymentService(
+	logger *zap.Logger,
+) *DeploymentService {
+	return &DeploymentService{
+		RegexPool: regexpool.DefaultPool(),
+		Logger:    logger.Named("planner"),
+	}
+}
+
+// Register registers all handling of this service into the specified gRPC server.
+func (a *DeploymentService) Register(server *grpc.Server) {
+	deployment.RegisterDeploymentServiceServer(server, a)
+}
+
+// DetermineStrategy implements deployment.DeploymentServiceServer.
+func (a *DeploymentService) DetermineStrategy(context.Context, *deployment.DetermineStrategyRequest) (*deployment.DetermineStrategyResponse, error) {
+	panic("unimplemented")
+}
+
+// DetermineVersions implements deployment.DeploymentServiceServer.
+func (a *DeploymentService) DetermineVersions(context.Context, *deployment.DetermineVersionsRequest) (*deployment.DetermineVersionsResponse, error) {
+	panic("unimplemented")
+}
+
+// BuildPipelineSyncStages implements deployment.DeploymentServiceServer.
+func (a *DeploymentService) BuildPipelineSyncStages(context.Context, *deployment.BuildPipelineSyncStagesRequest) (*deployment.BuildPipelineSyncStagesResponse, error) {
+	panic("unimplemented")
+}
+
+// BuildQuickSyncStages implements deployment.DeploymentServiceServer.
+func (a *DeploymentService) BuildQuickSyncStages(context.Context, *deployment.BuildQuickSyncStagesRequest) (*deployment.BuildQuickSyncStagesResponse, error) {
+	panic("unimplemented")
+}
+
+// FetchDefinedStages implements deployment.DeploymentServiceServer.
+func (a *DeploymentService) FetchDefinedStages(context.Context, *deployment.FetchDefinedStagesRequest) (*deployment.FetchDefinedStagesResponse, error) {
+	panic("unimplemented")
+}
